requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asteroid==0.7.0
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
boto3==1.35.70
botocore==1.35.70
cached-property==2.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
einops==0.8.0
elevenlabs==1.13.0
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mir_eval==0.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.20.1
openai==1.57.2
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pb-bss-eval==0.0.2
pesq==0.0.4
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
pyparsing==3.2.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytorch-ranger==0.1.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.12.0
srt==3.5.3
starlette==0.41.3
sympy==1.13.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.2.2
torch-optimizer==0.1.0
torch-stoi==0.2.3
torchaudio==2.2.2
torchmetrics==0.11.4
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vosk==0.3.44
websockets==14.1
whisper==1.1.10
yarl==1.18.3
